requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.3.0
pytesseract==0.3.10
pdf2image==1.17.0
pydantic==2.6.4
python-dotenv==1.0.1
google-generativeai==0.3.2